Add constants.py and data_viz.py
- moved global variables from av_etl.py to constants.py
- moved create_fig, add_subplots and visualize_data functions
from av_etl.py to data_viz.py
