+++ aws-lambda/Dockerfile
@@ -8,6 +8,8 @@ RUN  pip3 install -r requirements.txt
 
 # Copy function code
 COPY av_etl.py ${LAMBDA_TASK_ROOT}
+COPY constants.py ${LAMBDA_TASK_ROOT}
+COPY data_viz.py ${LAMBDA_TASK_ROOT}
 COPY aws-lambda/app.py ${LAMBDA_TASK_ROOT}
 
 # Set the CMD to your handler (could also be done as a parameter override outside of the Dockerfile)
